+++ test/coverage/test_phase.py
@@ -1895,6 +1895,36 @@ def test_missing_authorization(ctx, cli, snapshot_cli, openapi3_base_url):
     )
 
 
+def test_unnecessary_auth_warning(ctx, cli, snapshot_cli, openapi3_base_url):
+    # If a test for missing Authorization is the only thing that happen, there should be no warning for missing Authorization header
+    schema_path = ctx.openapi.write_schema(
+        {
+            "/basic": {
+                "get": {
+                    "security": [{"Basic": None}],
+                    "responses": {
+                        "200": {
+                            "description": "Ok",
+                        }
+                    },
+                }
+            }
+        },
+        version="2.0",
+        securityDefinitions={"Basic": {"type": "basic", "name": "Authorization", "in": "header"}},
+    )
+    assert (
+        cli.main(
+            "run",
+            str(schema_path),
+            f"--url={openapi3_base_url}",
+            "--header=Authorization: Basic dGVzdDp0ZXN0",
+            "--max-examples=5",
+        )
+        == snapshot_cli
+    )
+
+
 @pytest.mark.openapi_version("3.0")
 def test_nested_parameters(ctx):
     schema = ctx.openapi.build_schema(